requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
argcomplete==2.0.0
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
base58==2.1.1
bitarray==2.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cytoolz==0.12.1
Django==4.1.6
django-mathfilters==1.0.0
dnspython==2.3.0
eth-abi==4.0.0b2
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.0
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.2.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
packaging==23.0
parsimonious==0.9.0
Pillow==9.5.0
pipx==1.1.0
pluggy==1.0.0
protobuf==4.21.12
py-solc-x==1.1.1
pycodestyle==2.10.0
pycryptodome==3.16.0
pymongo==4.3.3
pyrsistent==0.19.3
pytest==7.2.1
python-dotenv==0.21.1
redis==4.5.4
regex==2022.10.31
requests==2.28.2
rlp==3.0.0
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
toolz==0.12.0
typing-extensions==3.10.0.2
urllib3==1.26.14
userpath==1.8.0
varint==1.0.2
web3 @ git+https://github.com/ethereum/web3.py.git@96b7785cfee9c71a72ce40981004a887543ae860
websockets==10.4
yarl==1.8.2
